certs: renew reload; check rtun.yaml before import

+++ nextbox_daemon/raw_backup_restore.py
@@ -276,7 +276,9 @@ class RawBackupRestore:
 
     def import_nextbox_dir(self, src_path):
         # copy rtun.yaml
-        shutil.copy(src_path / "nextbox" / self.nextbox_rtun, self.dirs["nextbox"])
+        p = src_path / "nextbox" / self.nextbox_rtun
+        if p.exists():
+            shutil.copy(src_path / "nextbox" / self.nextbox_rtun, self.dirs["nextbox"])
 
         # copy nextbox.conf
         shutil.copy(src_path / "nextbox" / self.nextbox_conf, self.dirs["nextbox"])
